+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
  - rename: `FileInfo` to `File`
  - new   : `File` => `getDir` > returns directory, file's return parent directory
+ - new   : `File` => `read` > gets file contents
+ - new   : `File` => `write` > sets file contents
  - update: `constructor` if no argument supplied the current directory is used
  - update: `getFile` & `getFiles` return `File` object(s)
  - minor : Improvements & fixes